Rename middleware drivers package.
